+++ include/layered_hardware_gazebo/posvel_pid_mode.hpp
@@ -40,8 +40,9 @@ public:
     const double pos(Position(*joint_, 0));
     // the latest position may be useful in the starting procedure of a pos-based controller
     data->position = pos;
-    pos_cmd_ = pos;
-    vel_lim_ = 0.;
+    pos_goal_ = pos;
+    pos_sp_ = pos;
+    vel_prof_ = 0.;
 
     pid_.reset();
   }
@@ -53,18 +54,20 @@ public:
   }
 
   virtual void write(const ti::RawJointData &data) {
-    pos_cmd_ = data.position_cmd;
-    vel_lim_ = std::abs(data.velocity_cmd);
+    pos_goal_ = data.position_cmd;
+    vel_prof_ = std::abs(data.velocity_cmd);
   }
 
   virtual void update(const ros::Time &time, const ros::Duration &period) {
-    // velocity required to realize the desired position in the next simulation step
-    const double vel_max((pos_cmd_ - Position(*joint_, 0)) / period.toSec());
-    // clamp the required velocity with the limits
-    const double vel_cmd(boost::algorithm::clamp(vel_max, -vel_lim_, vel_lim_));
-    const double vel_err(vel_cmd - joint_->GetVelocity(0));
-    const double eff_cmd(
-        boost::algorithm::clamp(pid_.computeCommand(vel_err, period), -eff_lim_, eff_lim_));
+    namespace ba = boost::algorithm;
+
+    // update the position setpoint
+    const double dpos_sp_lim(vel_prof_ * period.toSec());
+    pos_sp_ += ba::clamp(pos_goal_ - pos_sp_, -dpos_sp_lim, dpos_sp_lim);
+
+    // track the position setpoint
+    const double pos_err(pos_sp_ - Position(*joint_, 0));
+    const double eff_cmd(ba::clamp(pid_.computeCommand(pos_err, period), -eff_lim_, eff_lim_));
     if (!boost::math::isnan(eff_cmd)) {
       joint_->SetForce(0, eff_cmd);
     }
@@ -78,7 +81,7 @@ public:
 
 private:
   const double eff_lim_;
-  double pos_cmd_, vel_lim_;
+  double pos_goal_, pos_sp_, vel_prof_;
   control_toolbox::Pid pid_;
 };
 } // namespace layered_hardware_gazebo